-added TLSFAllocator utility class -removed temporary staging buffer -VKMemoryAllocator now uses TLSFAllocator class -fixed size_t -> uint32_t and similar warnings -added function to get number of allocations of an ObjectPool

+++ VEngine/Code/Engine.cpp
@@ -39,7 +39,7 @@ void VEngine::Engine::start()
 	while (!m_shutdown && !m_window->shouldClose())
 	{
 		timer.update();
-		double timeDelta = timer.getTimeDelta();
+		float timeDelta = static_cast<float>(timer.getTimeDelta());
 
 		m_window->pollEvents();
 		m_userInput->input();
